fix get owner func

+++ src/postgres/create.sql
@@ -59,7 +59,7 @@ create table if not exists service (
   merchant_id uuid not null references merchant(id),
   category_id int not null references service_category(id),
   name varchar(64) not null,
-  price int not null,
+  price number(12, 2) not null,
   image_url varchar(256),
   is_active boolean not null default false,
   deleted boolean not null default false
